CITYAI/requirements.txt
fastapi
uvicorn
python-multipart>=0.0.9
PyPDF2
pdfplumber>=0.11
camelot-py[cv]
tabula-py
pytesseract
opencv-python
pillow
scikit-learn>=1.3
matplotlib>=3.7
openpyxl>=3.1
pandas>=2.2
python-ternary
